==> 11-Data_preprocess.ipynb <==
import os
import docx
import aspose.words as aw
import jpype
import asposecells
jpype.startJVM()


from asposecells.api import Workbook, FileFormatType,SaveFormat

path = '/home/guoy20/Data'
classifications = ['Good','Bad','Inter']

finenames_good[:5]

def convert_to_json(file_path,file_name,classification):
    '''
    convert docx file to json file
    '''
    #Document docx file
    doc = aw.Document(file_path + '/' + classification + '/'+ file_name )
    
    #get the name without .docx to name the new json file
    new_file_name = file_name[:-5]
    
    #html path
    html_path = '/home/guoy20/Data/html' + '/' + classification + '/' + new_file_name + '.html'
    
    #convert to html file
    doc.save(html_path, aw.SaveFormat.HTML)
    
    #json_path
    json_path = '/home/guoy20/Data/json' + '/' + classification + '/' + new_file_name + '.json'
    
    # convert html to json
    book = Workbook(html_path)
    book.save(json_path, SaveFormat.JSON)
    
                      

#go through all the docx
for classification in classifications:
    finenames = os.listdir(path + '/' + classification)
    for filename in finenames:
        convert_to_json('/home/guoy20/Data',filename,classification)
