==> 22-Feature-generation.ipynb <==
# 22-Feature-generation

# to get more relationship between whether the employeee will be success or not

'''
feature including

words frequency: sales, investment, data, 

number density

document length

positive emotion score
negative emotion score

two other different emotion scores


'''

import pandas as pd
import os
import json
from collections import Counter
import xgboost as xgb
from sklearn.metrics import classification_report
from sklearn.model_selection import train_test_split
import torch
import readability

feature_data = pd.DataFrame()

file_path = []
file_name = []
home_path = '/home/guoy20/Data/json/'
label = []
classification = ['Good','Bad','Inter']
for c in classification:
    path1 = home_path + c 
    filename = os.listdir(path1)
    for name in filename:
        if name != '.ipynb_checkpoints':
            file_path.append(path1 + '/' + name)

            if c == 'Good':
                label.append(2)
            elif c == 'Inter':
                label.append(1)
            elif c == 'Bad':
                label.append(0)
            file_name.append(name)

def count_word(w,file_path):
    result = []
    for file in file_path:
        try:
            with open(file) as f:
                data = json.load(f)
        except:
            result.append(0)
            continue
        word_count = 0
        for item in data:

            for i in w:
                try:
                    word_count += list(item.values())[0].count(i)
                except:
                    word_count = word_count
        result.append(word_count)
    return result

sale_count = count_word(['sales','sale','Sales','Sale'],file_path = file_path)
financial_count = count_word(['financial','Financial'],file_path = file_path)
investment_count = count_word(['investment','Investment'],file_path = file_path)
advisor_count = count_word(['advisor','Advisor','advisors','Advisors'],file_path = file_path)

feature_data['label'] = label
feature_data['resume_name'] = file_name
feature_data['sales_count'] = sale_count
feature_data['financial_count'] = financial_count
feature_data['investment_count'] = investment_count
feature_data['advisor_count'] = advisor_count

feature_data

from transformers import AutoTokenizer, AutoModelWithLMHead

tokenizer = AutoTokenizer.from_pretrained("mrm8488/t5-base-finetuned-emotion")

model = AutoModelWithLMHead.from_pretrained("mrm8488/t5-base-finetuned-emotion")

def get_emotion(text):
    input_ids = tokenizer.encode(str(text) + '</s>', return_tensors='pt')
    output = model.generate(input_ids=input_ids,
               max_length=2)
    dec = [tokenizer.decode(ids) for ids in output]
    label = dec[0]
    return label
get_emotion("") # Output: 'joy'
 
get_emotion("i have a feeling i kinda lost my best friend")

final = []
for file in file_path:
    try:
        with open(file) as f:
            data = json.load(f)
    except:
        final.append({'joy':0,'fear':0,'sadness':0,'anger':0,'love':0,'surprise':0})
        continue
    document = {'joy':0,'fear':0,'sadness':0,'anger':0,'love':0,'surprise':0}
    count = 0
    for item in data:
        count += 1
        try:
            text = list(item.values())[0]
        except:
            continue
            
        
        
        result = get_emotion(text)[6:]
        if result in list(document.keys()):
            document[result] += 1
    r = {}
    for key in document:
        r[key] = document[key]/count
    final.append(r)
    

joy = [i['joy'] for i in final]
anger = [i['anger'] for i in final]
fear = [i['fear'] for i in final]

feature_data['joy'] = joy
feature_data['anger'] = anger 
feature_data['fear'] = fear

feature_data

number_count = []
for file in file_path:
    try:
        with open(file) as f:
            data = json.load(f)
    except:
        number_count.append(0)
        continue
    l = []
    for item in data:
        
        try:
            sentence = list(item.values())[0]
    
            if ('$' in sentence or '%' in sentence):
                t = sentence.split()
                for word in t:
                    if '$' in word or '%' in sentence:
                        l.append(word)
        except:
            continue
    number_count.append(len(l))
    

feature_data['number_count'] = number_count

from transformers import AutoTokenizer, DistilBertForSequenceClassification
tokenizer = AutoTokenizer.from_pretrained('has-abi/distilBERT-finetuned-resumes-sections')

model = DistilBertForSequenceClassification.from_pretrained("has-abi/distilBERT-finetuned-resumes-sections")

labels = ['awards','certificates','contact/name/title','education','interests','languages','para','professional_experiences','projects','skills','soft_skills','summary']

final_result = []
for file in file_path:
    d = {'awards':0,'certificates':0,'contact/name/title':0,'education':0,'interests':0,'languages':0,'para':0,'professional_experiences':0,'projects':0,'skills':0,'soft_skills':0,'summary':0}
    count = 0
    try:
        with open(file) as user_file:
            file_contents = user_file.read()
        list_file = eval(file_contents)
    
        for line in list_file:
            text = list(line.values())[0]
            encoded_input = tokenizer(text, return_tensors='pt')
            output = model(**encoded_input)
            label_id = torch.argmax(output.logits)
            label_id  = label_id.tolist()
            d[labels[label_id]] += 1
            count += 1


        for key in d:
            d[key]= d[key]/count
        final_result.append(d)
    except:
        final_result.append(d)
        continue

for i in labels:
    list1 = [j[i] for j in final_result]
    print(len(list1))
    if len(list1) == 130:
        feature_data[i] =list1

feature_data

l = []
final_r = []
for file in file_path:
    try:
        with open(file) as user_file:
            file_contents = user_file.read()
        list_file = eval(file_contents)
        for line in list_file:
            text = list(line.values())[0]
            results = readability.getmeasures(text, lang='en')
            score = results['readability grades']['FleschReadingEase']
            l.append(score)
        r = sum(l)/len(l)
        final_r.append(r)
            
    except:
        final_r.append(0)
        continue

feature_data['readability'] = final_r

feature_data

final_data = feature_data.sample(frac=1, random_state=42)

final_data

final_data.to_csv('/home/guoy20/Data/feature_data.csv', index=False)
