==> 23-new-emotion-analysis.ipynb <==
from transformers import AutoModelForSequenceClassification
from transformers import TFAutoModelForSequenceClassification
from transformers import AutoTokenizer
import numpy as np
from scipy.special import softmax
import csv
import urllib.request
import os




task='sentiment'
MODEL = f"cardiffnlp/twitter-roberta-base-{task}"

tokenizer = AutoTokenizer.from_pretrained(MODEL)
# download label mapping
labels=[]
mapping_link = f"https://raw.githubusercontent.com/cardiffnlp/tweeteval/main/datasets/{task}/mapping.txt"
with urllib.request.urlopen(mapping_link) as f:
    html = f.read().decode('utf-8').split("\n")
    csvreader = csv.reader(html, delimiter='\t')
labels = [row[1] for row in csvreader if len(row) > 1]


# PT
model = AutoModelForSequenceClassification.from_pretrained(MODEL)
model.save_pretrained(MODEL)



text = "Good night 😊"

encoded_input = tokenizer(text, return_tensors='pt')
output = model(**encoded_input)
scores = output[0][0].detach().numpy()
scores = softmax(scores)

ranking = np.argsort(scores)
ranking = ranking[::-1]
for i in range(scores.shape[0]):
    l = labels[ranking[i]]
    s = scores[ranking[i]]
    print(f"{i+1}) {l} {np.round(float(s), 4)}")


scores

classification = ['Good','Bad','Inter']

home_path = '/home/guoy20/Data/json/'
result = {'Good':[],'Bad':[],'Inter':[]}


for c in classification:
    class_path = home_path + c
    filenames = os.listdir(class_path)
    c_class = []
    for file in filenames:
        file_path = class_path +'/' + file
        try:
            with open(file_path) as user_file:
                file_contents = user_file.read()
            list_file = eval(file_contents)
            for line in list_file:
                text = list(line.values())[0]
                encoded_input = tokenizer(text, return_tensors='pt')
                output = model(**encoded_input)
                scores = output[0][0].detach().numpy()
                scores = softmax(scores)
                ranking = np.argsort(scores)
                ranking = ranking[::-1]
                l = labels[ranking[0]]
                s = scores[ranking[0]]
                f = {l:s}                         
            c_class.append(f)
            
        except:
            continue
        result[c] = c_class

result
