==> 25_emotion_exploration.ipynb <==
import pandas as pd

path = '/workspace/deerproject2/Data/feature_data.csv'

feature_data = pd.read_csv(path)

feature_data

def rank_quntile(df,name):
    new_df = df
    new_df['rank'] = new_df[name].rank(ascending=False)
    new_df['frac_rank'] = new_df['rank'] / len(df)
    new_df['quintile'] = pd.qcut(new_df['frac_rank'], q=5, labels=[1,2,3,4,5])
    grouped = new_df.groupby('quintile')['label'].value_counts().unstack()
    grouped['Total'] = (grouped[2] + grouped[0]+ grouped[1])
    grouped['Good/Bad+Good'] = (grouped[2])/(grouped[2] + grouped[0])
    grouped['Good/Total'] = (grouped[2])/(grouped['Total'])
    return grouped

rank_quntile(feature_data,'joy')

rank_quntile(feature_data,'fear')

rank_quntile(feature_data,'anger')

rank_quntile(feature_data,'number_count')

rank_quntile(feature_data,'professional_experiences')

feature_data.professional_experiences
