==> 32-transformer.ipynb <==
import torch
import torch.nn as nn
from transformers import DistilBertTokenizer, DistilBertForSequenceClassification, AdamW
from torch.utils.data import DataLoader, RandomSampler, SequentialSampler
from sklearn.metrics import accuracy_score
import pandas as pd
import os
import json
import unicodedata

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

device 

tokenizer = DistilBertTokenizer.from_pretrained('distilbert-base-uncased')

feature_data = pd.DataFrame()

file_path = []
file_name = []
home_path = '/home/guoy20/Data/json/'
label = []
classification = ['Good','Bad','Inter']
for c in classification:
    path1 = home_path + c 
    filename = os.listdir(path1)
    for name in filename:
        if name != '.ipynb_checkpoints':
            file_path.append(path1 + '/' + name)

            if c == 'Good':
                label.append(2)
            elif c == 'Inter':
                label.append(1)
            elif c == 'Bad':
                label.append(0)
            file_name.append(name)

dataset = []
for p in file_path:
    text = ''
    try:
        with open(p) as f:
            data = json.load(f)
        j = 0
        for item in data:
        
            try:
                text = text + list(item.values())[0]
                
            except:
                text = text
            j += 1
            if j > 5:
                break
    except:
        text = text
    new_str = unicodedata.normalize("NFKD", text)
    dataset.append(new_str)

dataset[0]

from transformers import AutoTokenizer, AutoModelForSequenceClassification, Trainer, TrainingArguments

model_name = "bert-base-uncased"
num_labels = 2 # Number of classes
model = AutoModelForSequenceClassification.from_pretrained(model_name, num_labels=num_labels)

# Load the tokenizer
tokenizer = AutoTokenizer.from_pretrained(model_name)


len(label) == len(dataset)

train_texts = dataset
train_labels = label# Class labels for each document

# Encode the training data
train_encodings = tokenizer(train_texts, truncation=True, padding=True)

class DocumentDataset(torch.utils.data.Dataset):
    def __init__(self, encodings, labels):
        self.encodings = encodings
        self.labels = labels

    def __getitem__(self, idx):
        item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
        item['labels'] = torch.tensor(self.labels[idx])
        return item

    def __len__(self):
        return len(self.labels)

train_dataset = DocumentDataset(train_encodings, train_labels)

# Define the training arguments
training_args = TrainingArguments(
    output_dir='./results',          # output directory
    num_train_epochs=3,              # total number of training epochs
    per_device_train_batch_size=16,  # batch size per device during training
    per_device_eval_batch_size=64,   # batch size for evaluation
    warmup_steps=500,                # number of warmup steps for learning rate scheduler
    weight_decay=0.01,               # strength of weight decay
    logging_dir='./logs',            # directory for storing logs
)

# Define the Trainer
trainer = Trainer(
    model=model,                         # the PyTorch model to be trained
    args=training_args,                  # training arguments, defined above
    train_dataset=train_dataset,         # training dataset
)

# Train the model
trainer.train()
