==> 33-random-forest.ipynb <==
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
import pandas as pd
from sklearn.metrics import classification_report

feature_data = pd.read_csv('/home/guoy20/Data/feature_data.csv')

# Split the data into training and testing sets
X = feature_data[list(feature_data.columns[2:])]
y = feature_data['label']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)

model = RandomForestClassifier(n_estimators=100, random_state=42)
model.fit(X_train, y_train)

y_pred = model.predict(X_test)

report = classification_report(y_test, y_pred)
print(report)

importance = model.feature_importances_
result = []

for i, v in enumerate(importance):
    result.append((X.columns[i], v))
result.sort(key = lambda x:x[1],reverse = True)
result

new_data = feature_data.drop(feature_data[(feature_data['label'] == 1)].index)

new_data.loc[new_data['label']==2, 'label'] = 1

X = new_data[list(new_data.columns[2:])]
y = new_data['label']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)

model = RandomForestClassifier(n_estimators=100, random_state=42)
model.fit(X_train, y_train)

y_pred = model.predict(X_test)

report = classification_report(y_test, y_pred)
print(report)

importance = model.feature_importances_
result = []

for i, v in enumerate(importance):
    result.append((X.columns[i], v))
result.sort(key = lambda x:x[1],reverse = True)
result
